api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
motor>=3.5.0
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.7.0
httpx>=0.27.0
openai>=1.40.0
python-dotenv>=1.0.1
email-validator>=2.2.0
aiosmtplib>=3.0.1
pywebpush>=1.14.0
APScheduler>=3.10.4
